refactor: simplify package synchronization
