+++ src/commands/sync.rs
@@ -1,10 +1,10 @@
 use crate::{
-    SyncError,
+    cli::SyncArgs,
     output::status,
     project::{
         LoadProjectResolutionError, ProjectLoadError, load_project, load_project_resolution,
     },
-    sync::{ProjectPackageMode, SyncProjectOptions, SystemSyncMode, sync_resolved_project},
+    sync::{SyncError, sync_resolved_project},
 };
 use miette::Diagnostic;
 use thiserror::Error;
@@ -24,32 +24,10 @@ pub(crate) enum Error {
     Sync(#[from] SyncError),
 }
 
-pub(crate) async fn run(
-    no_install_project: bool,
-    install_system: bool,
-    install_only_system: bool,
-) -> Result<(), Error> {
+pub(crate) async fn run(args: SyncArgs) -> Result<(), Error> {
     let project = load_project()?;
     let resolution = load_project_resolution(&project).await?;
-    sync_resolved_project(
-        &project,
-        resolution,
-        SyncProjectOptions {
-            project_package: if no_install_project {
-                ProjectPackageMode::Omit
-            } else {
-                ProjectPackageMode::Install
-            },
-            system: if install_only_system {
-                SystemSyncMode::InstallOnly
-            } else if install_system {
-                SystemSyncMode::Install
-            } else {
-                SystemSyncMode::Check
-            },
-        },
-    )
-    .await?;
+    sync_resolved_project(&project, resolution, args.no_install_project.into()).await?;
     status("Synchronized project library");
     Ok(())
 }
